Move `guess_elements()` to public API
