+++ doc/apidoc.json
@@ -287,7 +287,8 @@
         "Repair" : [
             "renumber_atom_ids",
             "renumber_res_ids",
-            "create_continuous_res_ids"
+            "create_continuous_res_ids",
+            "infer_elements"
         ],
         "Residue level utility" : [
             "get_residue_starts",
